+++ .eslintrc.js
@@ -5,5 +5,7 @@ module.exports = {
   root: true,
   rules: {
     "@typescript-eslint/ban-types": "off",
+    "@typescript-eslint/no-non-null-assertion": "warning",
+    "@typescript-eslint/no-var-requires": "warning",
   },
 }