+++ test/utf8.spec.ts
@@ -1,5 +1,5 @@
 import assert from "assert"
-import { utf8Decode } from "../src/decoder/utf8-decode.js"
+import { utf8Decode } from "../src/decoder/text-decode"
 
 // https://zh.wikipedia.org/wiki/%E7%B6%AD%E5%9F%BA%E7%99%BE%E7%A7%91%E6%A8%99%E8%AA%8C
 const textCases = [